+++ CHANGELOG.rst
@@ -8,6 +8,8 @@ Changelog
 
 Release 17/05/2023
 
+- Remove :code:`to_subcatalog` and :code:`to_subcatalog_dict` methods (:issue:`31`, :pull:`32`). 
+  By `Dougie Squire <https://github.com/dougiesquire>`_.
 - Re-implement the way that matched iterables are concatenated in searches to avoid having 
   to determine the iterable type from the first element (:issue:`29`, :pull:`30`). By 
   `Dougie Squire <https://github.com/dougiesquire>`_.